# docs/openapi.yaml
openapi: 3.0.3
info:
  title: AgentVerse Mission Engine API
  version: 1.0.0
  description: Authenticated Go backend for AI-native map missions, wallet-guarded LLM actions, profiles, and legacy CaseMind compatibility.
servers:
  - url: http://localhost:8080
security:
  - bearerAuth: []
...
paths:
...
  /api/v1/auth/register:
    post:
      tags: [Auth]
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/RegisterRequest" }
      responses:
        "201": { description: Registered }
        "400": { description: Invalid input }
  /api/v1/auth/login:
    post:
      tags: [Auth]
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/LoginRequest" }
      responses:
        "200": { description: Logged in }
        "401": { description: Invalid credentials }
  /api/v1/auth/refresh:
    post:
      tags: [Auth]
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [refresh_token]
              properties:
                refresh_token: { type: string }
      responses:
        "200": { description: Token refreshed }
  /api/v1/auth/logout:
    post:
      tags: [Auth]
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [refresh_token]
              properties:
                refresh_token: { type: string }
      responses:
        "200": { description: Logged out }
...
  /api/v1/profile:
    get:
      tags: [Profile]
      responses:
        "200":
          description: Player profile
          content:
            application/json:
              schema:
                type: object
                properties:
                  profile: { $ref: "#/components/schemas/Profile" }
    put:
      tags: [Profile]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [display_name]
              properties:
                display_name: { type: string, maxLength: 60 }
      responses:
        "200": { description: Updated profile }
...
  /api/v1/wallet/transactions:
    get:
      tags: [Wallet]
      parameters:
        - in: query
          name: limit
          schema: { type: integer, default: 50, maximum: 200 }
      responses:
        "200": { description: Wallet transactions }
...
  /api/v1/wallet/rewarded-ad/claim:
    post:
      tags: [Wallet]
      summary: Claim the daily rewarded-ad bonus (demo-gated)
      description: >
        Credits the fixed rewarded-ad payout, enforcing a server-side daily
        cap. Disabled (409 demo_disabled) when demo purchases are off.
      responses:
        "200": { description: Reward credited }
        "409": { description: Daily limit reached or demo disabled }
  /api/v1/wallet/purchase/verify:
    post:
      tags: [Wallet]
      summary: Verify a store purchase (demo/mock, gated in production)
      description: >
        Records the receipt (deduplicated) and credits the product coins.
        Store-side receipt verification is a placeholder; disabled
        (409 demo_disabled) when demo purchases are off (default in production).
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [platform, product_id, receipt]
              properties:
                platform: { type: string, enum: [ios, android] }
                product_id: { type: string, enum: [coins_small, coins_medium, coins_large] }
                receipt: { type: string }
      responses:
        "200": { description: Purchase credited }
        "409": { description: Receipt already used }
...
  /api/v1/missions:
    post:
      tags: [Missions]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/CreateMissionRequest" }
      responses:
        "202": { description: Mission accepted for generation }
        "409": { description: Insufficient wallet balance }
    get:
      tags: [Missions]
      responses:
        "200": { description: User missions }
  /api/v1/missions/{missionID}:
    get:
      tags: [Missions]
      parameters: [ { $ref: "#/components/parameters/MissionID" } ]
      responses:
        "200": { description: Mission dashboard }
        "404": { description: Not found or not owned }
  /api/v1/missions/{missionID}/dashboard:
    get:
      tags: [Missions]
      summary: Structured mission dashboard
      description: >
        Objectives (primary/required/optional/final), deterministic mission
        progress and risk scores, time remaining, next recommended actions,
        win/loss hints, annotated map markers, and completion readiness.
      parameters: [ { $ref: "#/components/parameters/MissionID" } ]
      responses:
        "200":
          description: Structured mission state
          content:
            application/json:
              schema: { $ref: "#/components/schemas/MissionDashboard" }
        "404": { description: Not found or not owned }
  /api/v1/missions/{missionID}/gameplay-status:
    get:
      tags: [Missions]
      summary: Single HUD payload with stages, clue goal, suspect status, board art
      description: >
        Superset of the dashboard: mission stages (with derived requirement
        progress), current stage, mission-wide clue goal, suspect status
        (hidden/identified), next stage reward, report-pending flag for the
        glowing report CTA, and generated board art. Runs a lazy, idempotent
        stage-engine pass so transitions earned by any action surface here.
      parameters: [ { $ref: "#/components/parameters/MissionID" } ]
      responses:
        "200":
          description: Gameplay status
          content:
            application/json:
              schema: { $ref: "#/components/schemas/GameplayStatus" }
        "404": { description: Not found or not owned }
  /api/v1/missions/{missionID}/actions/preview:
    post:
      tags: [Missions]
      summary: Preview an action's time/coin cost and risk before running it
      parameters: [ { $ref: "#/components/parameters/MissionID" } ]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [action]
              properties:
                action:
                  type: string
                  enum: [travel, location_action, character_chat, clue_inspect, report_submit]
                target_id: { type: string, format: uuid }
      responses:
        "200":
          description: Deterministic action preview (no state change)
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ActionPreview" }
        "400": { description: Unknown action }
  /api/v1/missions/{missionID}/reports:
    get:
      tags: [Reports]
      summary: Mission report history
      parameters: [ { $ref: "#/components/parameters/MissionID" } ]
      responses:
        "200":
          description: Reports, newest first
          content:
            application/json:
              schema:
                type: object
                properties:
                  reports:
                    type: array
                    items: { $ref: "#/components/schemas/MissionReport" }
    post:
      tags: [Reports]
      summary: Submit a report to command
      description: >
        Deterministic backend validation — the AI never mutates state and
        hidden truth is never consulted. Accepted reports advance stages,
        unlock content, and cost mission time; rejections explain exactly
        which requirements are missing. Types: clue_report, suspect_report,
        progress_report, incident_report, final_report.
      parameters: [ { $ref: "#/components/parameters/MissionID" } ]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [type, summary]
              properties:
                type:
                  type: string
                  enum: [clue_report, suspect_report, progress_report, incident_report, final_report]
                title: { type: string, maxLength: 200 }
                summary: { type: string, maxLength: 4000 }
                linked_clue_ids:
                  type: array
                  items: { type: string, format: uuid }
                suspect_character_id: { type: string, format: uuid }
      responses:
        "200":
          description: Verdict plus every state change the submission caused
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ReportResult" }
        "409": { description: Mission not active }
  /api/v1/missions/{missionID}/art/board/generate:
    post:
      tags: [Missions]
      summary: Generate scenario board background art
      description: >
        Renders one board background (mission_board_background,
        map_board_background, report_center_background, debrief_background,
        loading_screen) from player-safe world state — mission type, region,
        time of day, weather, risk mood — with HUD-safe empty areas and no
        text/logos. Generation prompts never leave the server; a procedural
        themed gradient is the guaranteed fallback.
      parameters: [ { $ref: "#/components/parameters/MissionID" } ]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [board_type]
              properties:
                board_type:
                  type: string
                  enum: [mission_board_background, map_board_background, report_center_background, debrief_background, loading_screen]
      responses:
        "200":
          description: Generated board (URL + status + version only)
          content:
            application/json:
              schema:
                type: object
                properties:
                  board: { $ref: "#/components/schemas/BoardImage" }
  /api/v1/missions/{missionID}/completion-check:
    get:
      tags: [Missions]
      summary: Whether the mission can be completed yet
      description: Returns can_complete and, when false, the missing requirements. Never charged.
      parameters: [ { $ref: "#/components/parameters/MissionID" } ]
      responses:
        "200":
          description: Completion readiness
          content:
            application/json:
              schema: { $ref: "#/components/schemas/CompletionCheck" }
  /api/v1/missions/{missionID}/complete:
    post:
      tags: [Missions]
      summary: Submit the final mission decision
      description: >
        Gates on readiness (returns 409 with missing requirements when not
        ready), otherwise runs the JudgeAgent and returns the win/loss result
        with objective- and clue-level explanation, stars, XP, and coins.
      parameters: [ { $ref: "#/components/parameters/MissionID" } ]
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                outcome: { type: string }
                reasoning: { type: string }
      responses:
        "200":
          description: Judged mission result
          content:
            application/json:
              schema: { $ref: "#/components/schemas/MissionResult" }
        "409": { description: Not ready to complete — missing requirements }
  /api/v1/missions/{missionID}/result:
    get:
      tags: [Missions]
      summary: Persisted result of a finished mission
      description: >
        Returns the stored end-of-mission result so it can be reviewed after
        the completion modal closes and from history. 409 when the mission has
        not reached a terminal state; 404 when no result was stored.
      parameters: [ { $ref: "#/components/parameters/MissionID" } ]
      responses:
        "200":
          description: Stored mission result
          content:
            application/json:
              schema:
                type: object
                properties:
                  mission_id: { type: string, format: uuid }
                  mission_status: { type: string, enum: [completed, failed] }
                  result: { $ref: "#/components/schemas/MissionResult" }
        "404": { description: No result stored }
        "409": { description: Mission not finished yet }
  /api/v1/missions/{missionID}/archive:
    post:
      tags: [Missions]
      parameters: [ { $ref: "#/components/parameters/MissionID" } ]
      responses:
        "200": { description: Mission archived }
  /api/v1/missions/{missionID}/events:
    get:
      tags: [Missions]
      summary: Raw chronological mission event feed (dev/detail)
      description: >
        Raw mission events, newest first. Clients building the player-facing
        story should prefer GET /missions/{missionID}/timeline.
      parameters:
        - { $ref: "#/components/parameters/MissionID" }
        - in: query
          name: limit
          schema: { type: integer, default: 100, maximum: 200 }
      responses:
        "200": { description: Mission event log }
  /api/v1/missions/{missionID}/timeline:
    get:
      tags: [Missions]
      summary: Curated player-facing mission timeline
      description: >
        The official story timeline: normalized event categories
        (mission_started, location_visited, clue_discovered, character_talked,
        ai_guidance_received, time_advanced, objective_completed, ...),
        oldest first, built only from player-safe event payloads. Internal
        noise (generation progress, raw facts, pacing hints) never appears.
      parameters:
        - { $ref: "#/components/parameters/MissionID" }
        - in: query
          name: limit
          schema: { type: integer, default: 200, maximum: 200 }
      responses:
        "200":
          description: Curated timeline
          content:
            application/json:
              schema: { $ref: "#/components/schemas/TimelineView" }
        "404": { description: Not found or not owned }
  /api/v1/missions/{missionID}/stream:
    get:
      tags: [Missions]
      parameters: [ { $ref: "#/components/parameters/MissionID" } ]
      responses:
        "200": { description: Server-sent mission events }
  /api/v1/missions/{missionID}/map:
    get:
      tags: [Map]
      parameters: [ { $ref: "#/components/parameters/MissionID" } ]
      responses:
        "200":
          description: Google Maps-compatible markers
          content:
            application/json:
              schema: { $ref: "#/components/schemas/MapView" }
  /api/v1/missions/{missionID}/locations/{locationID}:
    get:
      tags: [Map]
      parameters: [ { $ref: "#/components/parameters/MissionID" }, { $ref: "#/components/parameters/LocationID" } ]
      responses:
        "200": { description: Location bottom-sheet detail }
        "404": { description: Hidden or missing location }
  /api/v1/missions/{missionID}/locations/{locationID}/actions:
    post:
      tags: [Map]
      parameters: [ { $ref: "#/components/parameters/MissionID" }, { $ref: "#/components/parameters/LocationID" } ]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [action]
              properties:
                action: { type: string, example: inspect_area }
      responses:
        "200": { description: Action result and discovered clues }
  /api/v1/missions/{missionID}/locations/{locationID}/ask-ai:
    post:
      tags: [Guidance]
      parameters: [ { $ref: "#/components/parameters/MissionID" }, { $ref: "#/components/parameters/LocationID" } ]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [message]
              properties:
                message: { type: string }
      responses:
        "200": { description: Location-scoped AI guidance }
  /api/v1/missions/{missionID}/characters:
    get:
      tags: [Characters]
      parameters: [ { $ref: "#/components/parameters/MissionID" } ]
      responses:
        "200":
          description: Public NPC list
          content:
            application/json:
              schema:
                type: object
                properties:
                  characters:
                    type: array
                    items: { $ref: "#/components/schemas/PublicCharacter" }
  /api/v1/missions/{missionID}/characters/{characterID}:
    get:
      tags: [Characters]
      parameters: [ { $ref: "#/components/parameters/MissionID" }, { $ref: "#/components/parameters/CharacterID" } ]
      responses:
        "200": { description: Public NPC detail and transcript }
  /api/v1/missions/{missionID}/characters/{characterID}/chat:
    post:
      tags: [Characters]
      parameters: [ { $ref: "#/components/parameters/MissionID" }, { $ref: "#/components/parameters/CharacterID" } ]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [message]
              properties:
                message: { type: string }
                location_id: { type: string, format: uuid }
      responses:
        "200":
          description: NPC reply, state deltas, and cost
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ChatResult" }
  /api/v1/missions/{missionID}/characters/{characterID}/avatar:
    post:
      tags: [Characters]
      summary: Generate and persist a character avatar
      description: >
        Renders an avatar (image API or procedural fallback), stores it as a
        data URL, and returns the public character with avatar_url/avatar_status
        set. On generation failure avatar_status becomes "unavailable" so the
        client shows a placeholder instead of blocking. Never returns prompts.
      parameters: [ { $ref: "#/components/parameters/MissionID" }, { $ref: "#/components/parameters/CharacterID" } ]
      responses:
        "200":
          description: Character with avatar wired
          content:
            application/json:
              schema:
                type: object
                properties:
                  character: { $ref: "#/components/schemas/PublicCharacter" }
  /api/v1/missions/{missionID}/clues/{clueID}/image:
    post:
      tags: [Clues]
      summary: Generate and persist a clue evidence image
      description: >
        Renders an evidence image, stores it as a data URL, and returns the
        public clue with image_url/image_status set. Never returns prompts.
      parameters: [ { $ref: "#/components/parameters/MissionID" }, { $ref: "#/components/parameters/ClueID" } ]
      responses:
        "200":
          description: Clue with image wired
          content:
            application/json:
              schema:
                type: object
                properties:
                  clue: { $ref: "#/components/schemas/PublicClue" }
  /api/v1/missions/{missionID}/clues/{clueID}/confirm:
    post:
      tags: [Clues]
      summary: Confirm a discovered clue as evidence
      description: >
        Promotes a discovered/inspected clue to confirmed (evidence lifecycle),
        emits an evidence_confirmed timeline event, and runs the deterministic
        unlock engine — a confirmation may open the next locked location.
        Idempotent: confirming an already-confirmed clue does not unlock again.
        Free (a player decision, not an AI call). No hidden truth is exposed.
      parameters: [ { $ref: "#/components/parameters/MissionID" }, { $ref: "#/components/parameters/ClueID" } ]
      responses:
        "200":
          description: Progression envelope
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ProgressionEnvelope" }
  /api/v1/missions/{missionID}/hypothesis:
    post:
      tags: [Missions]
      summary: Submit a mid-mission working theory (checkpoint)
      description: >
        Returns a deterministic verdict from confirmed-evidence coverage
        (too_early | unsupported | partially_correct) plus feedback and next
        actions. Never claims correct/wrong (that is the final decision) and
        never reveals hidden truth. The AI does not mutate state here.
      parameters: [ { $ref: "#/components/parameters/MissionID" } ]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [answer]
              properties:
                answer: { type: string, maxLength: 2000 }
                clue_ids:
                  type: array
                  items: { type: string, format: uuid }
      responses:
        "200":
          description: Hypothesis checkpoint result
          content:
            application/json:
              schema:
                type: object
                properties:
                  verdict: { type: string, enum: [too_early, unsupported, partially_correct] }
                  feedback: { type: string }
                  needs_more_evidence: { type: boolean }
                  confirmed_count: { type: integer }
                  progression: { $ref: "#/components/schemas/ProgressionEnvelope" }
  /api/v1/missions/{missionID}/clues:
    get:
      tags: [Clues]
      parameters: [ { $ref: "#/components/parameters/MissionID" } ]
      responses:
        "200":
          description: Discovered clues only
          content:
            application/json:
              schema:
                type: object
                properties:
                  clues:
                    type: array
                    items: { $ref: "#/components/schemas/PublicClue" }
  /api/v1/missions/{missionID}/clues/{clueID}:
    get:
      tags: [Clues]
      parameters: [ { $ref: "#/components/parameters/MissionID" }, { $ref: "#/components/parameters/ClueID" } ]
      responses:
        "200": { description: Public clue detail }
...
  /api/v1/missions/{missionID}/time:
    get:
      tags: [Time]
      parameters: [ { $ref: "#/components/parameters/MissionID" } ]
      responses:
        "200": { description: Mission clock and public state }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    MissionID: { in: path, name: missionID, required: true, schema: { type: string, format: uuid } }
...
  schemas:
    RegisterRequest:
      type: object
      required: [email, password, display_name]
      properties:
        email: { type: string, format: email }
        password: { type: string, minLength: 8, maxLength: 72 }
        display_name: { type: string, maxLength: 60 }
    LoginRequest:
      type: object
      required: [email, password]
      properties:
        email: { type: string, format: email }
        password: { type: string }
...
    Profile:
      type: object
      properties:
        id: { type: string, format: uuid }
        user_id: { type: string, format: uuid }
        display_name: { type: string }
        rank: { type: string }
        level: { type: integer }
        xp: { type: integer }
        total_missions: { type: integer }
        completed_missions: { type: integer }
        failed_missions: { type: integer }
        success_rate: { type: number }
        favorite_mission_type: { type: string }
        total_clues_found: { type: integer }
        total_ai_interactions: { type: integer }
        total_locations_visited: { type: integer }
        badges: {}
        created_at: { type: string, format: date-time }
        updated_at: { type: string, format: date-time }
...